+++ fabricio/tasks.py
@@ -284,8 +284,11 @@ class DockerTasks(Tasks):
         """
         if self.registry is None and self.account is None:
             return
+        image = self.image[tag]
+        if not image:
+            return
         fabricio.local(
-            'docker pull {image}'.format(image=self.image[tag]),
+            'docker pull {image}'.format(image=image),
             quiet=False,
             use_cache=True,
         )
@@ -312,17 +315,20 @@ class DockerTasks(Tasks):
         """
         if self.registry is None and self.account is None:
             return
-        tag_with_registry = str(self.image[self.registry:tag:self.account])
+        image = self.image[tag]
+        if not image:
+            return
+        proxy_tag = image[self.registry:tag:self.account]
         fabricio.local(
             'docker tag {image} {tag}'.format(
-                image=self.image[tag],
-                tag=tag_with_registry,
+                image=image,
+                tag=proxy_tag,
             ),
             use_cache=True,
         )
         self.push_image(tag=tag)
         fabricio.local(
-            'docker rmi {tag}'.format(tag=tag_with_registry),
+            'docker rmi {tag}'.format(tag=proxy_tag),
             use_cache=True,
         )
 
@@ -444,22 +450,22 @@ class ImageBuildDockerTasks(DockerTasks):
         """
         build Docker image (see 'docker build --help' for available options)
         """
-        # default options
-        kwargs.setdefault('pull', True)
-        kwargs.setdefault('force-rm', True)
-
         for key, value in kwargs.items():
             try:
                 kwargs[key] = utils.strtobool(value)
             except ValueError:
                 pass
 
-        options = utils.Options(
-            tag=self.image[self.registry:tag:self.account],
-            **kwargs
-        )
+        image = self.image[self.registry:tag:self.account]
+        options = utils.Options(tag=image, **kwargs)
+
+        # default options
+        options.setdefault('pull', 1)
+        options.setdefault('force-rm', 1)
+
         fabricio.local(
             'docker build {options} {build_path}'.format(
+                image=image,
                 build_path=self.build_path,
                 options=options,
             ),